refactor rpt2_geoclaw.f, cleaned up and modified behavior
The logic is now cleaner and some work is skipped in dry cells so it
is faster.  Relative to earlier tests on this branch, this goes back
to the original approach of putting fluctuations in bmasdq if s<0 and
into bpasdq if s>0 without splitting if s near 0.  But it now skips
the splitting entirely only if eta < min(topo1,topo3) rather than max(..)
and creates transverse fluctuation going into the cell if eta < topo,
which seems more correct but could cause changes in results.