Add a temporary testing workflow
It uses PyClaw's tests since, to my understanding, an ensemble
of local regression tests for the Riemann repository using
Pytest would require the appropriate argument intents of the
rp* subroutines to be explicitly declared already.
So the Riemann solvers should be modernized first.

Also, the test for shallow water on the sphere is omitted for the moment.