Update and improve Riemann solver code (#176)
* Start using implicit none in all Riemann solvers.

* Start using implicit none in all Riemann solvers.

* Add a temporary testing workflow

It uses PyClaw's tests since, to my understanding, an ensemble
of local regression tests for the Riemann repository using
Pytest would require the appropriate argument intents of the
rp* subroutines to be explicitly declared already.
So the Riemann solvers should be modernized first.

Also, the test for shallow water on the sphere is omitted for the moment.

* Fix typo branch name

* Use implicit none in more Riemann solvers

* implicit none in SWEs on the sphere

The PyClaw test that uses this RS is skipped in the testing
workflow since it leads to a core dump
(I suspect this has to do with the OS or the system
architecture).

However, this test passes locally for me (before and after the
implicit none changes).

* All 2D RSs using implicit none

* implicit none used in every Riemann solver

* Add workflow for Classic tests

* Temporary fix Pytest not collecting AMRClaw tests

The regression tests in AMRClaw are not being collected by Pytest. They
just have to be renamed. This is a temporary fix to run them from this
repo in a single pytest session.

---------

Co-authored-by: Carlos Munoz Moncayo <carlos.munozmoncayo@kaust.edu.sa>